+++ genetic_algorithms/algorithm_wrappers/algorithm_wrapper.py
@@ -14,31 +14,13 @@ class AlgorithmWrapper(NextStateProvider):
         self.algorithm = algorithm
 
     @abstractmethod
-    def _perform_side_effects(self, model: Model, state: State):
+    def _perform_side_effects(self, model: 'Model', state: 'State'):
         """
         Performs logic related to the plugin.
         """
 
-    def next_state(self, model: Model, state: State) -> Union[State, None]:
+    def next_state(self, model: 'Model', state: 'State') -> Union['State', None]:
         next_state = self.algorithm.next_state(model, state)
         if next_state:
             self._perform_side_effects(model=model, state=next_state)
         return next_state
-
-
-class VisualizationWrapper(AlgorithmWrapper):
-    """
-    Provides visualization to algorithm solutions.
-    """
-    visualizations = {}
-
-    def __init__subclass__(cls):
-        VisualizationWrapper.visualizations[cls.get_corresponding_problem(
-        )] = cls
-
-    @staticmethod
-    @abstractmethod
-    def get_corresponding_problem() -> Type[Model]:
-        """
-        Returns model type for which this visualization was done
-        """
