+++ genetic_algorithms/algorithm_wrappers/visualizations/__init__.py
@@ -1,0 +1,2 @@
+from genetic_algorithms.algorithm_wrappers.visualizations.visualization_wrapper import VisualizationWrapper
+from genetic_algorithms.algorithm_wrappers.visualizations.traveling_salesman_visualization import TravelingSalesmanVisualization
