+++ genetic_algorithms/models/next_state_provider.py
@@ -1,12 +1,12 @@
 from abc import ABC, abstractmethod
 from typing import Union
-from genetic_algorithms.problems.base.state import State
-from genetic_algorithms.problems.base.model import Model
+from genetic_algorithms.problems.base import State
+# from genetic_algorithms.problems.base.model import Model
 
 
 class NextStateProvider(ABC):
     @abstractmethod
-    def next_state(self, model: Model, state: State) -> Union[State, None]:
+    def next_state(self, model: 'Model', state: 'State') -> Union['State', None]:
         """
         Returns next state that could be reaced from the current state.
         """
