+++ genetic_algorithms/problems/__init__.py
@@ -1,3 +1,3 @@
-from genetic_algorithms.problems.base import Model
+from genetic_algorithms.problems.base import Model, State
 from genetic_algorithms.problems.traveling_salesman_problem.problem_model import TravelingSalesmanProblem
 from genetic_algorithms.problems.graph_coloring_problem.problem_model import GraphColoringProblem
