+++ run.py
@@ -1,0 +1,4 @@
+from genetic_algorithms.cli.entry_point import entry_point
+# In case you are a Windows user :)
+if __name__ == '__main__':
+    entry_point()
